Extend QbeastSnapshot with new methods
